Initial task servers configuration

+++ main.py
@@ -1,17 +1,48 @@
 from __future__ import annotations
-from importlib import import_module
+
+import os
 import socket
+from importlib import import_module
+from time import sleep, time
+
+from paste.translogger import TransLogger
 from waitress import serve
-from task_runner.app import app
-from task_runner.logger import logger
+
 import task_runner.routes  # pylint: disable=unused-import
-from task_runner.args import task_name, reload_enabled
-from utils.paths.create_join_paths_fn import (create_join_paths_fn)
+from task_runner.app import app
+from task_runner.args import reload_enabled, task_name
+from task_runner.logger import logger
+from utils.constants import APP_TASK_NAME_KEY
+from utils.paths.create_join_paths_fn import create_join_paths_fn
 from utils.paths.get_dirname import get_dirname
 
 PORT_RANGE = [7100, 7199]
 
 dirname = get_dirname(__file__)
+
+
+# pylint: disable=redefined-outer-name
+def is_port_open(port: int, timeout=10) -> bool:
+    """
+    Checks if port is open given a timeout.
+
+    Returns:
+    True if port is open before timeout else False.
+    """
+    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    timeout_end = timeout + time()
+
+    logger.info(f"Waiting for port {port} to open...")
+    while time() < timeout_end:
+        try:
+            sock.bind(('', port))
+            sock.close()
+            logger.info(f"Port {port} opened!")
+            return True
+        except OSError as e:
+            sleep(0.25)
+    logger.info(f"Port {port} did open after {timeout} seconds. Looking for a different port...")
+    return False
 
 
 # pylint: disable=redefined-outer-name
@@ -64,17 +95,44 @@ def get_allocated_task_ports() -> set[int]:
 # pylint: enable=redefined-outer-name
 
 if __name__ == "__main__":
-    allocated_task_ports = get_allocated_task_ports()
-    port = next_free_port(allocated_task_ports)
-    join_paths = create_join_paths_fn(dirname)
-    js_port_file_path = join_paths("tasks", task_name, "port.js", rm=True)
-    js_port_file_path.touch()
-    js_port_file_path.write_text(f"export default {port};\n")
+    # Set if not in the flask reloading child
+    # https://stackoverflow.com/questions/25504149/why-does-running-the-flask-dev-server-run-itself-twice
+    port = None
 
-    py_port_file_path = join_paths("tasks", task_name, "port.py", rm=True)
-    py_port_file_path.touch()
-    py_port_file_path.write_text(f"PORT = {port}\n")
+    app.config[APP_TASK_NAME_KEY] = task_name
 
+    is_reloading_child = os.environ.get("WERKZEUG_RUN_MAIN") == "true"
+    logger.info("is_reloading_child=%s", is_reloading_child)
+    # Only update port before debug mode has started or in production WSGI server
+    if not is_reloading_child:
+        join_paths = create_join_paths_fn(dirname)
+        py_port_file_path = join_paths("tasks", task_name, "port.py", rm=False)
+
+        # We need to use the same port because sometimes when restarting, the port
+        # is still occupied for the next minute for some reason. Waitress serve
+        # will always overwrite the port.
+        if py_port_file_path.exists():
+            port = getattr(import_module(f"tasks.{task_name}.port"), 'PORT')
+            logger.info("Using existing port: %s", port)
+        else:
+            logger.info("looking for port")
+
+            # if not existing_port_is_open:
+            allocated_task_ports = get_allocated_task_ports()
+            port = next_free_port(allocated_task_ports)
+
+            py_port_file_path = join_paths("tasks", task_name, "port.py", rm=True)
+            py_port_file_path.touch()
+            py_port_file_path.write_text(f"PORT = {port}\n")
+
+            js_port_file_path = join_paths("tasks", task_name, "port.js", rm=True)
+            js_port_file_path.touch()
+            js_port_file_path.write_text(f"export default {port};\n")
+
+        logger.info("App is started.")
+        logger.info(f"Listening on port {port}")
+
+    # app.config["started"] = True
     # Setting the app to serve single threaded because it will entirely
     # avoid the problems of multi-threading dealing
     # with file writes and reads that may interfere with one another
@@ -82,10 +140,10 @@ if __name__ == "__main__":
     # something to aim for in the future, then switching from files
     # to a network database should be the first step to take because it will
     # allow transaction locks and get the most out of peformance of reads and writes.
-
     if reload_enabled:
         logger.info("App reload is enabled.")
         # Reload
         app.run(debug=True, port=port, threaded=False)
     else:
-        serve(app, host="0.0.0.0", port=port, threads=1)
+        logger.info("App reload is disabled.")
+        serve(TransLogger(app), host="0.0.0.0", port=port, threads=1)
